Add check for currency values when using default from session Move context id to separate file to prevent circural deps
