+++ projects/core/src/occ/adapters/site-context/site-context.interceptor.ts
@@ -15,7 +15,7 @@ import { getContextParameterDefault } from '../../../site-context/config/context
 import {
   CURRENCY_CONTEXT_ID,
   LANGUAGE_CONTEXT_ID,
-} from '../../../site-context/providers/context-service-map';
+} from '../../../site-context/providers/context-ids';
 
 @Injectable()
 export class SiteContextInterceptor implements HttpInterceptor {
@@ -41,9 +41,9 @@ export class SiteContextInterceptor implements HttpInterceptor {
       .getActive()
       .subscribe(data => (this.activeLang = data));
 
-    this.currencyService
-      .getActive()
-      .subscribe(data => (this.activeCurr = data));
+    this.currencyService.getActive().subscribe(data => {
+      this.activeCurr = data;
+    });
   }
 
   intercept(
