+++ projects/core/src/site-context/config/base-site-config-validator.ts
@@ -1,6 +1,6 @@
 import { SiteContextConfig } from './site-context-config';
-import { BASE_SITE_CONTEXT_ID } from '../providers/context-service-map';
 import { getContextParameterDefault } from './context-config-utils';
+import { BASE_SITE_CONTEXT_ID } from '../providers/context-ids';
 
 export function baseSiteConfigValidator(config: SiteContextConfig) {
   if (getContextParameterDefault(config, BASE_SITE_CONTEXT_ID) === undefined) {
