+++ projects/core/src/site-context/providers/context-ids.ts
@@ -1,0 +1,3 @@
+export const LANGUAGE_CONTEXT_ID = 'language';
+export const CURRENCY_CONTEXT_ID = 'currency';
+export const BASE_SITE_CONTEXT_ID = 'baseSite';
