+++ projects/core/src/site-context/providers/context-service-map.ts
@@ -3,14 +3,15 @@ import { LanguageService } from '../facade/language.service';
 import { CurrencyService } from '../facade/currency.service';
 import { SiteContext } from '../facade/site-context.interface';
 import { BaseSiteService } from '../facade/base-site.service';
+import {
+  BASE_SITE_CONTEXT_ID,
+  CURRENCY_CONTEXT_ID,
+  LANGUAGE_CONTEXT_ID,
+} from './context-ids';
 
 export abstract class ContextServiceMap {
   [context: string]: Type<SiteContext<any>>;
 }
-
-export const LANGUAGE_CONTEXT_ID = 'language';
-export const CURRENCY_CONTEXT_ID = 'currency';
-export const BASE_SITE_CONTEXT_ID = 'baseSite';
 
 export function serviceMapFactory() {
   return {
