+++ projects/core/src/site-context/providers/index.ts
@@ -1,3 +1,4 @@
 export * from './context-service-map';
 export * from './context-service-providers';
 export * from './site-context-params-providers';
+export * from './context-ids';
